+++ src/commands/releases.rs
@@ -108,6 +108,10 @@ pub fn make_app<'a, 'b: 'a>(app: App<'a, 'b>) -> App<'a, 'b> {
                 .help("When the flag is set and the previous release commit was not found in the repository, \
                         will create a release with the default commits count (or the one specified with `--initial-depth`) \
                         instead of failing the command."))
+            .arg(Arg::with_name("ignore-empty")
+                .long("ignore-empty")
+                .help("When the flag is set, command will not fail and just exit silently \
+                        if no new commits for a given release have been found."))
             .arg(Arg::with_name("local")
                 .conflicts_with_all(&["auto", "clear", "commits", ])
                 .long("local")
@@ -170,7 +174,17 @@ pub fn make_app<'a, 'b: 'a>(app: App<'a, 'b>) -> App<'a, 'b> {
             .arg(Arg::with_name("show_projects")
                 .short("P")
                 .long("show-projects")
-                .help("Display the Projects column")))
+                .help("Display the Projects column"))
+            .arg(Arg::with_name("raw")
+                .short("R")
+                .long("raw")
+                .help("Print raw, delimiter separated list of releases. [defaults to new line]"))
+            .arg(Arg::with_name("delimiter")
+                .short("D")
+                .long("delimiter")
+                .takes_value(true)
+                .requires("raw")
+                .help("Delimiter for the --raw flag")))
         .subcommand(App::new("info")
             .about("Print information about a release.")
             .version_arg(1)
@@ -526,7 +540,7 @@ fn execute_set_commits<'a>(
     // make sure the release exists if projects are given
     if let Ok(projects) = ctx.get_projects(matches) {
         ctx.api.new_release(
-            &org,
+            org,
             &NewRelease {
                 version: version.into(),
                 projects,
@@ -556,7 +570,7 @@ fn execute_set_commits<'a>(
             }
             table.print();
         }
-        ctx.api.set_release_refs(&org, version, heads)?;
+        ctx.api.set_release_refs(org, version, heads)?;
     } else {
         let default_count = matches
             .value_of("initial-depth")
@@ -588,7 +602,14 @@ fn execute_set_commits<'a>(
         let commits = generate_patch_set(&repo, commit_log, prev_commit, &parsed)?;
 
         if commits.is_empty() {
-            bail!("No commits found. Leaving release alone.");
+            // TODO(v2): Make it a default behavior on next major release instead?
+            let ignore_empty = matches.is_present("ignore-empty");
+            if ignore_empty {
+                println!("No commits found. Leaving release alone.");
+                return Ok(());
+            } else {
+                bail!("No commits found. Change commits range, initial depth or use --ignore-empty to allow empty patch sets.");
+            }
         }
 
         ctx.api.update_release(
@@ -658,6 +679,18 @@ fn execute_restore<'a>(ctx: &ReleaseContext<'_>, matches: &ArgMatches<'a>) -> Re
 fn execute_list<'a>(ctx: &ReleaseContext<'_>, matches: &ArgMatches<'a>) -> Result<(), Error> {
     let project = ctx.get_project_default().ok();
     let releases = ctx.api.list_releases(ctx.get_org()?, project.as_deref())?;
+
+    if matches.is_present("raw") {
+        let versions = releases
+            .iter()
+            .map(|release_info| release_info.version.clone())
+            .collect::<Vec<_>>()
+            .join(matches.value_of("delimiter").unwrap_or("\n"));
+
+        println!("{}", versions);
+        return Ok(());
+    }
+
     let mut table = Table::new();
     let title_row = table.title_row();
     title_row.add("Released").add("Version");
@@ -706,7 +739,7 @@ fn execute_info<'a>(ctx: &ReleaseContext<'_>, matches: &ArgMatches<'a>) -> Resul
     let version = matches.value_of("version").unwrap();
     let org = ctx.get_org()?;
     let project = ctx.get_project_default().ok();
-    let release = ctx.api.get_release(org, project.as_deref(), &version)?;
+    let release = ctx.api.get_release(org, project.as_deref(), version)?;
 
     // quiet mode just exists
     if matches.is_present("quiet") {
@@ -863,7 +896,7 @@ fn execute_files_upload<'a>(
         let ctx = &UploadContext {
             org,
             project: project.as_deref(),
-            release: &version,
+            release: version,
             dist,
             wait: matches.is_present("wait"),
         };
@@ -883,9 +916,9 @@ fn execute_files_upload<'a>(
         if let Some(artifact) = ctx.api.upload_release_file(
             org,
             project.as_deref(),
-            &version,
+            version,
             &FileContents::FromPath(path),
-            &name,
+            name,
             dist,
             Some(&headers[..]),
             ProgressBarMode::Request,
@@ -1063,7 +1096,7 @@ fn execute_files_upload_sourcemaps<'a>(
 
     // make sure the release exists
     let release = ctx.api.new_release(
-        &org,
+        org,
         &NewRelease {
             version: version.into(),
             projects: ctx.get_projects(matches)?,